requirements.txt
# RAKSHA-FORCE Backend Dependencies
# Install: pip install -r requirements.txt

# Web framework (Vercel serverless compatible)
fastapi==0.115.5
mangum==0.19.0            # ASGI adapter for AWS Lambda / Vercel

# Supabase client
supabase==2.10.0

# JWT validation
PyJWT==2.10.1

# HTTP client (used by supabase-py)
httpx==0.27.2

# Data validation (FastAPI uses this internally)
pydantic==2.10.3
pydantic-settings==2.7.0

# ASGI server (for local development only)
uvicorn[standard]==0.32.1
